requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.8.1
asteval==1.0.4
astunparse==1.6.3
attrs==24.2.0
boto3==1.35.26
botocore==1.35.26
Brotli==1.1.0
bw-migrations==0.2
bw2calc==1.8.2
bw2data==3.6.6
bw2io==0.8.12
bw2parameters==1.1.0
cached-property==1.5.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
constructive_geometries==1.0
country-converter==1.2
dataflows-tabulator==1.54.3
datapackage==1.15.4
diff-match-patch==20230430
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-import-export==4.1.1
django-unfold==0.39.0
djangorestframework==3.15.2
docopt==0.6.2
ecoinvent_interface==2.5
eight==1.0.1
et-xmlfile==1.1.0
fasteners==0.19
flexcache==0.3
flexparser==0.3.1
future==0.18.3
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
ijson==3.3.0
inflate64==1.0.0
isodate==0.6.1
jmespath==1.0.1
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
linear-tsv==1.1.0
llvmlite==0.43.0
lxml==5.3.0
mrio-common-metadata==0.2.1
multivolumefile==0.2.3
numba==0.60.0
numpy==2.0.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
peewee==3.17.6
pillow==10.4.0
Pint==0.24.3
platformdirs==4.3.6
prettytable==3.11.0
psutil==6.0.0
psycopg2-binary==2.9.9
py7zr==0.22.0
pybcj==1.0.2
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyecospold==4.0.0
pyppmd==1.1.0
PyPrind==2.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
pyxlsb==1.0.10
PyYAML==6.0.2
pyzstd==0.16.1
referencing==0.35.1
requests==2.32.3
rfc3986==2.0.0
rpds-py==0.20.0
s3transfer==0.10.2
scipy==1.14.1
six==1.16.0
sparse==0.15.4
SQLAlchemy==2.0.35
sqlparse==0.5.1
stats_arrays==0.7
tableschema==1.20.11
tablib==3.5.0
texttable==1.7.0
toolz==0.12.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
unfold==1.2.0
unicodecsv==0.14.1
Unidecode==1.3.8
urllib3==2.2.3
voluptuous==0.15.2
wcwidth==0.2.13
whitenoise==6.7.0
Whoosh==2.7.4
wrapt==1.16.0
wurst==0.4
xlrd==2.0.1
XlsxWriter==3.2.0
